requirements.txt
kafka-python
sqlalchemy
pandas
psycopg2-binary